+++ eslint.config.mjs
@@ -26,6 +26,7 @@ const eslintConfig = [
       "dist/**",
       "build/**",
       "plugins/**",
+      "next-env.d.ts",
     ],
   },
 ];